Phase 2 of #827: extract Alpine factory from connection_detail (#862)
Move the per-connection detail page's Alpine factory out of the
~645-LOC `connection_detail_scripts.go` Go-string sidecar into
`static/js/admin/components/connection_detail.js`, following the
convention codified in #828.

The factory `connectionDetail` owns:
- the manual sync trigger + 5-minute poll loop with the optimistic
  pending sync-log row in #sync-history-list,
- per-connection actions (pause/resume, delete, sync-interval override),
- per-account actions (rename / include-exclude),
- the header "remove" confirm-then-confirm UX (formerly an inline
  `x-data="{ confirming: false }"`).

Inline `onclick`/`onchange` handlers in templ were replaced with Alpine
`@click`/`@change` directives so the JS sits in a single module. The
two `connDetail*Input` raw-string helpers were deleted and the inputs
now live as native templ markup. The connection ID flows in via a
single `data-conn-id` attribute on the x-data root and is read in
`init()`. No `@templ.JSONScript` is needed — all per-action URLs are
derived from the connection ID.

Pure-Go templating helpers (badges, type labels, bar styles, etc.)
that lived in `connection_detail_scripts.go` alongside the JS body
moved to a new `connection_detail_helpers.go`. The "_scripts.go" file
is deleted entirely as the per-page convention requires.

Lint ceiling unchanged — `connection_detail.templ` had no Source-B
inline `<script>` block (only the Source-A Go-string sidecar that this
PR removes), so the literal-script max stays at 147 in
connections.templ:454-620.

Validated in the browser:
- Page renders (header, sync-health stats, sync timeline, info grid,
  accounts grid with display-name/exclude inputs, sync history).
- Manual Sync Now triggers the fetch + polling loop end-to-end and
  the new sync log row shows up.
- Display-name update persists across reloads.
- Sync-interval change shows the "Saved" inline status.
- Remove confirm-then-confirm flow flips the inline buttons.
- Console silent (only pre-existing form-field-id and
  apple-mobile-web-app-capable warnings).
